Update R profile (colorout)
library() was causing errors if the package was not installed.
Switched to require(), which sends a warning instead, and returns whether the package was loaded successfully or not.
Moved related functions into an IF block based on the result of require().
Also added more comments and ideas for other things to add someday.